Move DIFF_OPEN_SIDE from editorCommands to diffEditorCommands; rename command palette action to Open Active Diff Side
